Only count spot requests tagged for our ASG
This allows multiple tests to run concurrently. It also avoids flapping
tests: Old spot requests get deleted after some time, potentially changing
the total number of spot requests while the test is running.

Tests need to be adapted because the untagging of a spot request is now
counted as a delta of -1, not 0.